requirements.txt
# requirements.txt
# Dependencias Python para el backend
# Instalar en local: pip install -r requirements.txt
# Instalar en Docker: automático (Dockerfile lo ejecuta)

# ============================================
# Framework Web
# ============================================
fastapi==0.104.1              # Framework web async
uvicorn[standard]==0.24.0     # ASGI server (corre FastAPI)
pydantic==2.5.0               # Validación de datos con Python type hints
pydantic-settings==2.1.0      # Manejo de settings/config desde .env

# ============================================
# Base de datos
# ============================================
sqlalchemy==2.0.23            # ORM y query builder
psycopg2-binary==2.9.9        # Driver PostgreSQL (psycopg2)
alembic==1.13.1               # Migrations (versionado de schema)

# ============================================
# Cache y Queue
# ============================================
redis==5.0.1                  # Cliente Redis
httpx==0.25.2                 # HTTP client (async) para llamadas externas

# ============================================
# Autenticación y Seguridad
# ============================================
python-jose[cryptography]==3.3.0  # JWT tokens
passlib[bcrypt]==1.7.4            # Password hashing
cryptography==41.0.7              # Librerías cryptográficas

# ============================================
# Logging y Monitoreo
# ============================================
python-json-logger==2.0.7     # Logging estructurado en JSON

# ============================================
# Testing
# ============================================
pytest==7.4.3                 # Framework de testing
pytest-asyncio==0.21.1        # Tests async
pytest-cov==4.1.0             # Coverage reports
httpx==0.25.2                 # HTTP client para tests

# ============================================
# Utilidades
# ============================================
python-dotenv==1.0.0          # Cargar variables desde .env
requests==2.31.0              # HTTP requests (sync)

# ============================================
# Desarrollo local
# ============================================
black==23.12.1                # Formateador de código
flake8==6.1.0                 # Linter (análisis estático)
mypy==1.7.1                   # Type checking
isort==5.13.2                 # Organizador de imports